Handle call update with ANSWERED state from Telephony
getRingingCall() can get calls in RINGING and
ANSWERED state if the call was already answered.
Get calls only in RINGING state to be able to
answer calls that are not already answered.

Ignore call update from RINGING->ANSWERED. Treat
ANSWERED call as RINGING as BT is not aware of
ANSWERED state. This will make sure proper call
updates, SCO creation is done when multiple calls
exist on the phone.

CRs-Fixed: 2471073
Change-Id: Ia513289b290f6a4ea49bf91c5cf628ad46fcb98c